Moved commands folder structure.
