+++ src/commands/main/all-tts/StopCommand.js
