+++ src/commands/main/config/SetDefaultProviderCommand.js
@@ -1,16 +1,17 @@
 const { SlashCommand } = require('@greencoast/discord.js-extended');
 const { SlashCommandBuilder } = require('@discordjs/builders');
 const logger = require('@greencoast/logger');
-const TTSPlayer = require('../../classes/tts/TTSPlayer');
+const TTSPlayer = require('../../../classes/tts/TTSPlayer');
 
-class SetMyProviderCommand extends SlashCommand {
+class SetDefaultProviderCommand extends SlashCommand {
   constructor(client) {
     super(client, {
-      name: 'set_my_provider',
-      description: 'Sets the provider to be used by the say command for yourself.',
+      name: 'set_default_provider',
+      description: 'Sets the provider to be used by the say command for the server by default.',
       emoji: ':pencil2:',
       group: 'config',
       guildOnly: true,
+      userPermissions: ['MANAGE_GUILD'],
       dataBuilder: new SlashCommandBuilder()
         .addStringOption((input) => {
           return input
@@ -27,11 +28,11 @@ class SetMyProviderCommand extends SlashCommand {
     const providerName = interaction.options.getString('provider');
     const providerFriendlyName = TTSPlayer.PROVIDER_FRIENDLY_NAMES[providerName];
 
-    await this.client.ttsSettings.set(interaction.member, { provider: providerName });
+    await this.client.ttsSettings.set(interaction.guild, { provider: providerName });
 
-    logger.info(`User ${interaction.member.displayName} in ${interaction.guild.name} has changed their provider to ${providerName}.`);
-    return interaction.reply({ content: localizer.t('command.set.my.provider.success', { name: providerFriendlyName }), ephemeral: true });
+    logger.info(`${interaction.guild.name} has changed its default provider to ${providerName}.`);
+    return interaction.reply({ content: localizer.t('command.set.default.provider.success', { name: providerFriendlyName }) });
   }
 }
 
-module.exports = SetMyProviderCommand;
+module.exports = SetDefaultProviderCommand;
