+++ src/commands/main/google-tts/GoogleLangsCommand.js
@@ -1,9 +1,9 @@
 const { SlashCommand } = require('@greencoast/discord.js-extended');
 const { SlashCommandBuilder } = require('@discordjs/builders');
 const { MessageEmbed } = require('discord.js');
-const { MESSAGE_EMBED } = require('../../common/constants');
-const { splitContentForEmbedFields } = require('../../utils/embed');
-const languages = require('../../../provider-data/google_languages.json');
+const { MESSAGE_EMBED } = require('../../../common/constants');
+const { splitContentForEmbedFields } = require('../../../utils/embed');
+const languages = require('../../../../provider-data/google_languages.json');
 
 class GoogleLangsCommand extends SlashCommand {
   constructor(client) {
