+++ src/commands/optional/channel-tts/config/ChannelSettingsCommand.js
@@ -2,8 +2,8 @@ const { SlashCommand } = require('@greencoast/discord.js-extended');
 const { SlashCommandBuilder } = require('@discordjs/builders');
 const { MessageEmbed } = require('discord.js');
 const merge = require('deepmerge');
-const { MESSAGE_EMBED } = require('../../common/constants');
-const TTSPlayer = require('../../classes/tts/TTSPlayer');
+const { MESSAGE_EMBED } = require('../../../../common/constants');
+const TTSPlayer = require('../../../../classes/tts/TTSPlayer');
 
 class ChannelSettingsCommand extends SlashCommand {
   constructor(client) {
